+++ tests/notifications_python_client/test_notifications_api_client.py
@@ -175,7 +175,7 @@ def test_create_email_notification_with_personalisation(notifications_client, rm
     }
 
 
-def test_create_email_notification_with_unsubscribe_link(notifications_client, rmock):
+def test_create_email_notification_with_one_click_unsubscribe_url(notifications_client, rmock):
     endpoint = "{0}/v2/notifications/email".format(TEST_HOST)
     rmock.request("POST", endpoint, json={"status": "success"}, status_code=200)
 
@@ -183,14 +183,14 @@ def test_create_email_notification_with_unsubscribe_link(notifications_client, r
         email_address="to@example.com",
         template_id="456",
         personalisation={"name": "Namey"},
-        unsubscribe_link="https://unsubscribelink.com/unsubscribe",
+        one_click_unsubscribe_url="https://unsubscribelink.com/unsubscribe",
     )
 
     assert rmock.last_request.json() == {
         "template_id": "456",
         "email_address": "to@example.com",
         "personalisation": {"name": "Namey"},
-        "unsubscribe_link": "https://unsubscribelink.com/unsubscribe",
+        "one_click_unsubscribe_url": "https://unsubscribelink.com/unsubscribe",
     }